Move EpochStartConfig to a separate file
