+++ crates/sui-core/src/authority/authority_per_epoch_store.rs
@@ -36,6 +36,7 @@ use typed_store::rocks::{DBBatch, DBMap, DBOptions, MetricConf, TypedStoreError}
 use typed_store::traits::{TableSummary, TypedStoreDebug};
 
 use crate::authority::authority_notify_read::NotifyRead;
+use crate::authority::epoch_start_configuration::EpochStartConfiguration;
 use crate::authority::{AuthorityStore, CertTxGuard, ResolverWrapper, MAX_TX_RECOVERY_RETRY};
 use crate::checkpoints::{
     CheckpointCommitHeight, CheckpointServiceNotify, EpochStats, PendingCheckpoint,
@@ -60,10 +61,9 @@ use std::cmp::Ordering as CmpOrdering;
 use sui_adapter::adapter;
 use sui_protocol_config::{ProtocolConfig, ProtocolVersion};
 use sui_storage::mutex_table::MutexGuard;
-use sui_types::epoch_data::EpochData;
 use sui_types::message_envelope::TrustedEnvelope;
 use sui_types::messages_checkpoint::{
-    CertifiedCheckpointSummary, CheckpointContents, CheckpointDigest, CheckpointSequenceNumber,
+    CertifiedCheckpointSummary, CheckpointContents, CheckpointSequenceNumber,
     CheckpointSignatureMessage, CheckpointSummary, CheckpointTimestamp,
 };
 use sui_types::storage::{transaction_input_object_keys, ObjectKey, ParentSync};
@@ -154,69 +154,6 @@ pub struct AuthorityPerEpochStore {
 
     /// Execution state that has to restart at each epoch change
     execution_component: ExecutionComponents,
-}
-
-/// Parameters of the epoch fixed at epoch start.
-#[derive(Serialize, Deserialize, Debug, Eq, PartialEq)]
-pub struct EpochStartConfiguration {
-    system_state: EpochStartSystemState,
-    /// epoch_digest is defined as following
-    /// (1) For the genesis epoch it is set to 0
-    /// (2) For all other epochs it is a digest of the last checkpoint of a previous epoch
-    /// Note that this is in line with how epoch start timestamp is defined
-    epoch_digest: CheckpointDigest,
-}
-
-impl EpochStartConfiguration {
-    pub fn new(system_state: EpochStartSystemState, epoch_digest: CheckpointDigest) -> Self {
-        Self {
-            system_state,
-            epoch_digest,
-        }
-    }
-
-    pub fn new_for_testing() -> Self {
-        Self::new(
-            EpochStartSystemState::new_for_testing(),
-            CheckpointDigest::default(),
-        )
-    }
-
-    pub fn epoch_data(&self) -> EpochData {
-        EpochData::new(
-            self.epoch(),
-            self.epoch_start_timestamp_ms(),
-            self.epoch_digest(),
-        )
-    }
-
-    pub fn epoch_digest(&self) -> CheckpointDigest {
-        self.epoch_digest
-    }
-
-    pub fn epoch(&self) -> EpochId {
-        self.system_state.epoch
-    }
-
-    pub fn protocol_version(&self) -> ProtocolVersion {
-        ProtocolVersion::new(self.system_state.protocol_version)
-    }
-
-    pub fn reference_gas_price(&self) -> u64 {
-        self.system_state.reference_gas_price
-    }
-
-    pub fn safe_mode(&self) -> bool {
-        self.system_state.safe_mode
-    }
-
-    pub fn epoch_start_timestamp_ms(&self) -> u64 {
-        self.system_state.epoch_start_timestamp_ms
-    }
-
-    pub fn epoch_start_state(&self) -> &EpochStartSystemState {
-        &self.system_state
-    }
 }
 
 /// AuthorityEpochTables contains tables that contain data that is only valid within an epoch.
