+++ crates/sui-core/src/checkpoints/checkpoint_executor/tests.rs
@@ -7,13 +7,13 @@ use tempfile::tempdir;
 
 use std::{sync::Arc, time::Duration};
 
+use crate::authority::epoch_start_configuration::EpochStartConfiguration;
 use broadcast::{Receiver, Sender};
 use sui_protocol_config::SupportedProtocolVersions;
 use sui_types::committee::ProtocolVersion;
 use sui_types::messages_checkpoint::{ECMHLiveObjectSetDigest, VerifiedCheckpoint};
 use tokio::{sync::broadcast, time::timeout};
 
-use crate::authority::authority_per_epoch_store::EpochStartConfiguration;
 use crate::{
     authority::AuthorityState, checkpoints::CheckpointStore, state_accumulator::StateAccumulator,
 };
